+++ RevisionApp/MainView/Modifiers/Trophy Color/TrophyColorProvider.swift
@@ -8,11 +8,11 @@
 import SwiftUI
 
 protocol TrophyColorProvider {
-    func trophyColor(for percentage: Double) -> Color
+    func trophyColor(for percentage: Int) -> Color
 }
 
 class DefaultTrophyColorProvider: TrophyColorProvider {
-    func trophyColor(for percentage: Double) -> Color {
+    func trophyColor(for percentage: Int) -> Color {
         switch percentage {
         case 100:
             return Color("gold")
